unittests/issue105 <flush_threshold>
In unittests/issue105 the flush threshold value is set if provided
on the command line. As a result for values in the range [1, 100]
the test passes for the following values:

$ for i in {1..100}; do ./unittests/issue105 $i &>/dev/null && echo $i; done
1
3
5
15
25
38
39
40
41
42
43
44
45
46
47
48
49
75

The total number of mutating operations carried out on the database by
this test case is 75 (50 insertions followed by 25 deletions). The test
passes when the flush frequency is a divisor of that number, as well as
when no erase operations are flushed (though the flush frequency 50 is a
strange exception). When the flush frequency exceeds the operation
count, the test fails because of wrong item count (0 instead of 25), in
all other cases - because of encountering a deleted item during
iteration.

This demonstrates that the problem is related to the handling of flushed
operations.