In TxnOperation fully replaced flags with effect
As a result of this change the effect of every successful insertion operation
performed in a transaction is determined immediately. This somewhat
simplifies the key counting code, however may result in worse performance
for insertions that are allowed to overwrite or duplicate existing keys
(before this change btree was not being checked for existence of keys
inserted in that mode).